### 2017-10-14 #429 @whedon
[PRE REVIEW]: Parallel and Adaptive High-Dimensional Numerical Integration
**Submitting author:** @iljah (<a href="http://orcid.org/0000-0002-9542-5866">Ilja Honkonen</a>)
**Repository:** <a href="https://github.com/iljah/hdintegrator" target ="_blank">https://github.com/iljah/hdintegrator</a>
**Version:** 0.1
