### 2017-10-23 #429 @marksantcroos
@danielskatz I was offline for a week. Will do an assessment whether I'm qualified for review today or tomorrow.
