### 2017-10-25 #437 @whedon
[REVIEW]: Parallel and Adaptive High-Dimensional Numerical Integration
**Submitting author:** @iljah (<a href="http://orcid.org/0000-0002-9542-5866">Ilja Honkonen</a>)
**Repository:** <a href="https://github.com/iljah/hdintegrator" target ="_blank">https://github.com/iljah/hdintegrator</a>
**Version:** 0.1

- [x] **License:** Does the repository contain a plain-text LICENSE file with the contents of an [OSI approved](https://opensource.org/licenses/alphabetical) software license? [#4](https://github.com/iljah/hdintegrator/issues/4)

- [x] **Installation:** Does installation proceed as outlined in the documentation? [#5](https://github.com/iljah/hdintegrator/issues/5) [#9](https://github.com/iljah/hdintegrator/issues/9)

- [x] **A statement of need:** Do the authors clearly state what problems the software is designed to solve and who the target audience is? [#7](https://github.com/iljah/hdintegrator/issues/7)
- [x] **Installation instructions:** Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution. [#1](https://github.com/iljah/hdintegrator/issues/5)
- [x] **Functionality documentation:** Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)? [#8](https://github.com/iljah/hdintegrator/issues/8)
- [x] **Automated tests:** Are there automated tests or manual steps described so that the function of the software can be verified? [#6](https://github.com/iljah/hdintegrator/issues/6)
