### 2017-11-02 #437 @danielskatz
Some comments from Paul Zimmerman that may help in the review process:

> The full paper is not a long paper. The review is most of the software.
>
> Starting from the issue (https://github.com/openjournals/joss-reviews/issues/429),
the paper is in the repository as paper.md
(https://github.com/iljah/hdintegrator/blob/master/paper.md). A PDF will be
generated if/when accepted.
>
> And again, the review is more of the software than the paper, but does include
both. See the review info in the about page
(http://joss.theoj.org/about#reviewer_guidelines), and see an example of a review
in progress: https://github.com/openjournals/joss-reviews/issues/407
