### 2017-11-06 #437 @marksantcroos
Hi @danielskatz, sorry for the delay. I actually have a question. I played around initially, but ran into some issues that stopped me from being able to continue. I meant to read more about the process but postponed that ... Is the process meant to be incremental, or should I do it based on the status quo? E.g., if the installation stops me from continuing, is that then the end-result of the first review, or should I wait for the author to fix it? (Assuming I create issues during my review)
