### 2017-11-06 #437 @danielskatz
Don't think of this as a first review, but as an incremental process.  Report problems to the author.

- If brief, report them here

- If detailed, open an issue in the source repo and make a note here that you have done so with a link to the issue(s).

Then you are done until the author responds...
