### 2017-11-06 #437 @danielskatz
Some comments from Paul Zimmerman that might help the review process:

> Reviewer: Paul Zimmermann
>
> This review is on revision 97f76c6c224c2433e32c8d3d03f640188e63a290
> (git clone https://github.com/iljah/hdintegrator.git).
>
> I was told by the editor (Daniel Katz) that there would be both a paper and
> a software to review, but the paper (file paper.md) is very short: it only
> consists of 36 lines. This "paper" mentions that hdintegrator was developed
> because "none [FOSS packages for numerical integration] support parallel
> integration in arbitrary number of dimensions". It would make sense to at
> least mention in the paper some examples of integrals computed with
> hdintegrator, with timings and discussing the accuracy of the result,
> and to compare timings and accuracy to other software tools. Also the
> References part of the paper is empty.
>
> The README file says that mpi4py is needed, but it is not explained how to get
it (say in Debian or Ubuntu).
>
> I tried the example in README but could not make it work:
>
> $ mpiexec -n 5 ./main_mpi.py --integrator integrands/N-sphere.py --dimensions 2
> Couldn't import required submodule(s), did you clone them with:
>
> git clone --recursive https://github.com/iljah/hdintegrator.git
> Couldn't import required submodule(s), did you clone them with:
>
> git clone --recursive https://github.com/iljah/hdintegrator.git
> - - - -
> Primary job  terminated normally, but 1 process returned
> a non-zero exit code.. Per user-direction, the job has been aborted.
> - - - -
> - - - -
> mpiexec detected that one or more processes exited with non-zero status, thus causing
> the job to be terminated. The first process to do so was:
>
>  Process name: [[23694,1],0]
>  Exit code:    1
>
> My opinion is that this paper is not ready for publication, and the software
> is not ready either. The paper should include at least a few examples of
> multiple integration with timings, accuracy, and comparison to other
> multi-dimensional integrators. The software should include detailed
> instructions about pre-requisites, and better explanations in case of error.
>
> Moreover, on http://joss.theoj.org/about#reviewer_guidelines best practices
> are mentioned: documentation, tests, continuous integration, and a license.
> While I see the license (file COPYING), I could not find the documentation
> about how to define a function to be integrated, I did not see a "tests"
> directory, and I did not see any continuous integration link.
