### 2017-11-06 #437 @danielskatz
In general, Paul's comments are inputs to you and @marksantcroos.  I don't think these discussions that Paul mentions are needed in paper.md, but they probably are needed somewhere in the documentation/repository, in my opinion.

Please feel free to ignore the references comment - Paul is new to the JOSS system.  Your understanding is correct.
