### 2017-11-06 #437 @iljah
mpi4py appears to be available in Ubuntu repositories so it should be installable as described in Ubuntu documentation: https://help.ubuntu.com/stable/ubuntu-help/addremove-install.html I don't use Ubuntu myself but I'll mention this and add the required command line incantations to the INSTALL file.

Looks like the installation instructions forgot to mention that the PYTHONPATH environment variable must include the directory of the required submodules that git clone --recursive downloads, I'll see if I can work around that within the program itself. If I cannot, I'll add a mention of updating PYTHONPATH to the INSTALL file. In case of other technical problems I suggest opening an issue in the hdintegrator repo so that others might find them, and their solutions, more easily.

The README file describes how the main program communicates with the program that does evaluates the integrals but I'll add a concrete example of how one could define a function to be integrated.
