### 2017-11-07 37cfef4
Append required submodules to sys.path.
Prevents need for manipulating PYTHONPATH manually.
