### 2017-11-07 a8e1d76
Add mention of networkx to INSTALL file.
Also add mention of required packages in Fedora and Ubuntu.
