### 2017-11-07 40d31fb
Print exception if importing submodules fails.
