### 2017-11-07 #437 @iljah
The program should work now without requiring manual editing of PYTHONPATH and I also added the NetworkX package as a requirement that I had forgotten to mention. I couldn't get the program to work with a vanilla install of Ubuntu but this seems to be a known bug for over a year now that is still unresolved: https://bugs.launchpad.net/ubuntu/+source/mpi4py/+bug/1583432 There's a workaround in the bug report but I didn't test it.

I added a mention of the required packages' names and a note that are different versions available for different openmpi implementation and python versions.
