### 2017-11-21 iljah/hdintegrator#1 @marksantcroos
pypi package
Is there a reason not to provide a pypi package?
I generally consider that a good practice.
