### 2017-11-21 iljah/hdintegrator#2 @marksantcroos
use markdown for readme, install, etc
Please use markdown to increase the readability of the documentation.
