### 2017-11-21 iljah/hdintegrator#3 @marksantcroos
recommend virtual environments in installation procedure
Instead of having people mess up their system environment, I would suggest you document the installation using virtual environments (or anaconda's equivalent).
