### 2017-11-21 iljah/hdintegrator#5 @marksantcroos
installation is not really documented
The INSTALL provides some details about the software requirements, but doesn't actually specify how to install the hdintegrator software itself.
