### 2017-11-21 iljah/hdintegrator#8 @marksantcroos
usage documentation largely missing
Please provide information like:
- arguments
- options
- input formats
