### 2017-11-21 iljah/hdintegrator#4 @iljah
FSF [instructions](https://www.gnu.org/licenses/gpl-howto.html) say to use a file named COPYING, would a symlink to LICENSE work?
