### 2017-11-21 iljah/hdintegrator#5 @iljah
Aside from checking out the repository there's nothing to install if one is using e.g. a python integrand. For included C++ integrand the file [does](https://github.com/iljah/hdintegrator/blob/master/INSTALL#L20) mention compiling them with (gnu) make using the provided makefile. I'll add an explicit mention of potentially not requiring any installation when I change the files to markdown format.
