### 2017-11-21 iljah/hdintegrator#6 @iljah
Do you mean something else besides or in addition to [this](https://github.com/iljah/hdintegrator/blob/a8e1d763790582a12f41ad93f4f93e3e438768e3/INSTALL#L27)?
