### 2017-11-21 iljah/hdintegrator#4 @marksantcroos
I suspected so much, as COPYING sounds more opinionated than LICENSE.

Honestly I don't feel strong about it. In fact, [Github documentation](https://help.github.com/articles/licensing-a-repository/) actually suggests adding a `LICENSE.md` or `LICENSE.txt`, which is not strictly the same as the review template either.

@danielskatz Maybe something to change in the review template?
