### 2017-11-21 iljah/hdintegrator#5 @marksantcroos
That needs to be changed then. I don't want to use code directly from a checkout.
