### 2017-11-21 iljah/hdintegrator#6 @marksantcroos
Something more automated that returns True or False.
