### 2017-11-21 iljah/hdintegrator#4 @danielskatz
Given that I see in the instructions

>You should also include a copy of the license itself somewhere in the distribution of your program. All programs, whether they are released under the GPL or LGPL, should include the text version of the GPL. In GNU programs we conventionally put the license in a file called COPYING.

it seems clear that the use of copying is a convention, not a requirement, and having an additional file called LICENSE, either that's a copy of COPYING or a simlink, should be fine.

Re the JOSS review template: LICENSE vs LICENSE.txt is not much of a concern for JOSS.  But LICENSE.md would not be ok - we say that this should be plaintext...
