### 2017-11-21 #437 @danielskatz
thanks @marksantcroos and @iljah

For the record, the pass by @marksantcroos led to this set of issues in the code repo:

- https://github.com/iljah/hdintegrator/issues/1
- https://github.com/iljah/hdintegrator/issues/2
- https://github.com/iljah/hdintegrator/issues/3
- https://github.com/iljah/hdintegrator/issues/4
- https://github.com/iljah/hdintegrator/issues/5
- https://github.com/iljah/hdintegrator/issues/6
- https://github.com/iljah/hdintegrator/issues/7
- https://github.com/iljah/hdintegrator/issues/8
