### 2017-11-22 ef00d3d
Rename copying to license.
Fixes https://github.com/iljah/hdintegrator/issues/4
