### 2017-11-22 iljah/hdintegrator#4 @iljah
Fixed by https://github.com/iljah/hdintegrator/commit/ef00d3d8f04fac9405bb2feaf22e4ec70e0fc01a
