### 2017-11-22 iljah/hdintegrator#8 @iljah
Arguments and options are described by the program if run with the --help option. I/O format is described in the README but I'll improve the formatting to make that easier to find.
