+++ README.md
@@ -1,25 +1,27 @@
 # HDIntegrator
 
 High-Dimensional Integrator (HDIntegrator) is a program for parallel numerical
-integration of functions in arbitrary number of dimensions, see paper.md for a
+integration of functions in arbitrary number of dimensions, see [paper.md](paper.md) for a
 general overview.
+
 
 # Installation
 
-Installation and testing is detailed in the INSTALL file but at least Python 3
+Installation and testing is detailed in the [INSTALL.md](INSTALL.md) file but at least Python 3
 and an implementation of Message Passing Interface is required along with mpi4py
 and the NetworkX python package. Separate programs are used for evaluating
 integrals numerically and are provided in the integrands directory. These can
-have their own prerequisites as detailed in INSTALL file.
+have their own prerequisites as detailed in [INSTALL](INSTALL.md).
+
 
 # Usage
 
 To get a list of mandatory and optional arguments supported by the program run
-main_mpi.py with --help argument: `./main_mpi.py --help`.
+[main_mpi.py](main_mpi.py) with --help argument: `./main_mpi.py --help`.
 
 In general the program must be run with at least two MPI processes and must
-include a serial program to use as an integrand and the number of dimensions to
-integrate in:
+include a serial program to use as an integrand as well as the number of
+dimensions to integrate in:
 
     mpiexec -n 5 ./main_mpi.py --integrator integrands/N-sphere.py --dimensions 2
 
@@ -27,10 +29,13 @@ The result consists of one line with the integral's value, absolute error and
 fraction of volume relative to total volume in which the integrand failed to
 return a value.
 
+
+# Input and output formats
+
 To define your own integral you must write a program that will be called by
-main_mpi.py for evaluating the intergral within a given volume and give that
+main_mpi.py for evaluating that intergral within a given volume and give that
 program as an argument to main_mpi.py. For an example see e.g. the program in
-integrands/N-sphere.py which evaluates the integral for an N-dimensional sphere.
+[integrands/N-sphere.py](integrands/N-sphere.py) which evaluates the integral for an N-dimensional sphere.
 Communication between main_mpi.py and integrands is handled via standard input
 and output in ASCII format. Each line given to the integrand by main_mpi.py
 consists of floating point numbers separated by spaces:
