### 2017-11-22 iljah/hdintegrator#3 @iljah
I'll add instructions for installing required python libraries under home directory but at least on fedora all requirements are available from distribution repositories so I wouldn't exactly call installing them "messing up the system environment".
