### 2017-11-22 iljah/hdintegrator#3 @iljah
Fixed by https://github.com/iljah/hdintegrator/commit/9ced7419b692788a313c325ec9a54224f0d6824c
