### 2017-11-22 51b0ca9
Convert readme in integrand dir to markdown.
