### 2017-11-22 iljah/hdintegrator#2 @iljah
Fixed in https://github.com/iljah/hdintegrator/commit/51b0ca9ce19dcbcd129360a78a53a25ee7e88836 and previous commits.
