### 2017-11-22 iljah/hdintegrator#5 @iljah
I'll use https://github.com/iljah/hdintegrator/issues/1 to track installability.
