### 2017-11-22 iljah/hdintegrator#1 @iljah
If I understand correctly, pypi uses python distutils to do the work but the first problem is that distutils doesn't seem to support building C++ executables. After following the documentation for extensions it only builds a .o file and I haven't found a way to convince it to create an executable that would be installed somewhere. At this point I think pypi installation will have to exclude non-python integrands...
