### 2017-11-22 7e10d09
Add tests to Makefile.
