### 2017-11-22 371eef2
Mention automatic tests in install file.
