### 2017-11-22 iljah/hdintegrator#6 @iljah
Fixed in https://github.com/iljah/hdintegrator/commit/7e10d09cb233b20d33441bc66151881a39a99177. Pytest at least doesn't support testing mpi4py programs so I decided a simple addition to the makefile would suffice.
