### 2017-11-22 bc960d5
Remove test results on make clean.
