### 2017-11-22 a5825f6
Rename references to integrator to integrand in code and docs.
