### 2017-11-22 1746cb0
Mention direct pip installation in install file.
