### 2017-11-22 780c19d
Add more info on obtaining all integrands to install file.
