### 2017-11-22 iljah/hdintegrator#1 @iljah
Fixed by https://github.com/iljah/hdintegrator/commit/05efc047b61def32645545dcee763f750b4eb476 and later commits as best as possible I think. If you know of a way to get setuptools to compile and install executables along with python code feel free to reopen. The version obtained from pip includes integrand(s) implemented in python.
