### 2017-11-22 4d7fff1
Spelling fix in readme.
