### 2017-11-22 b16f6fd
Fix type in example command in readme and add example output.
