### 2017-11-22 iljah/hdintegrator#8 @iljah
The readme should now have basic usage instructions and be better formatted so the information is easier to find.
