### 2017-11-22 585d27c
Add a link to wikipedia on numerical integration to readme.
