### 2017-11-22 iljah/hdintegrator#7 @iljah
I added a link in readme file to wikipedia article on numerical integration which seems like a good general introduction on the topic: https://github.com/iljah/hdintegrator/commit/585d27c2f683e2fbfb3bd75a812101c8e6cd25a8#diff-04c6e90faac2675aa89e2176d2eec7d8. Let me know if this wasn't what you had in mind.
