### 2017-11-22 #437 @iljah
I think I managed to fix the identified issues, if not feel free to reopen or create new ones.
