### 2017-11-27 iljah/hdintegrator#2 @marksantcroos
Thanks! Hopefully that didn't feel like a chore but I hope you consider it an improvement for yourself too! :-)
