### 2017-11-27 iljah/hdintegrator#9 @marksantcroos
local library / include path location
Can you possibly change the Makefile and/or document how to pick up the installed GSL files?

I had to set these environment variables to get passed:
```
export CPATH=/opt/local/include
export LIBRARY_PATH=/opt/local/lib
```
