### 2017-11-27 iljah/hdintegrator#11 @marksantcroos
split out python vs non-python
Given the complexity of installing the non-python bits, I would suggest splitting the two up.

The first part of the documentation would then be roughly:
- setup the virtual env
- `pip install hidintegrator`
- the python usage examples, e.g. `mpiexec -n 4 hdintegrator.py --integrand integrands/N-sphere.py --dimensions 2` (pointing out that need those files obviously)

And then continue, "for increased performance, one can install C++ versions of the integrands ..."

That at least gets your users with something attainable.
