### 2017-11-27 iljah/hdintegrator#12 @marksantcroos
ld: library not found for -lboost_program_options
Which dependency is missing here?

```
$ make
g++ -std=c++14 -O3 -march=native -W -Wall -Wextra -Wpedantic integrands/burgers.cpp -o integrands/burgers_plain integrands/gsl/plain2.c -DMETHOD=1 -I integrands/gsl -lgsl -lgslcblas -lboost_program_options
clang: warning: treating 'c' input as 'c++' when in C++ mode, this behavior is deprecated [-Wdeprecated]
ld: library not found for -lboost_program_options
clang: error: linker command failed with exit code 1 (use -v to see invocation)
make: *** [integrands/burgers_plain] Error 1
```
