### 2017-11-27 #437 @marksantcroos
Managed to run some of the python-only examples now, so thats progress!
