### 2017-11-27 #437 @danielskatz
new issues:
- https://github.com/iljah/hdintegrator/issues/9
- https://github.com/iljah/hdintegrator/issues/10
- https://github.com/iljah/hdintegrator/issues/11
- https://github.com/iljah/hdintegrator/issues/12
