### 2017-11-27 #437 @marksantcroos
@danielskatz Ack. Note that I've been (trying) to maintain a list of issues in the checklist above. I think the system could be of more help to us with some reworking. Happy to share my thoughts there.
