### 2017-11-27 #437 @danielskatz
I see - I didn't notice that, as I was just scanning the checklist to see if the boxes were getting checked, not realizing that you might add notes to the items.

Your suggestions are very welcome.  Please either open an issue at https://github.com/openjournals/joss/issues or if you are feeling brave, you can edit code and make a PR in that repo as well.
