### 2017-11-28 3cf5b6d
Reduce required accuracy for tests.
Remove non-working and unused integrand.
