+++ Makefile
@@ -1,6 +1,11 @@
-MPIEXEC=mpiexec
-DIFF=diff
-TOUCH=touch
+PYTHON = python
+MPIEXEC = mpiexec
+DIFF = diff
+TOUCH = touch
+CXX = c++
+CPPFLAGS ?=
+CXXFLAGS ?= -std=c++14 -O3 -march=native -W -Wall -Wextra -Wpedantic
+LDFLAGS ?=
 
 PROGRAMS=integrands/failing \
 	integrands/maybe_failing \
@@ -18,7 +23,7 @@ PROGRAMS=integrands/failing \
 
 all: $(PROGRAMS)
 
-COMP = g++ -std=c++14 -O3 -march=native -W -Wall -Wextra -Wpedantic $< -o $@
+COMP = $(CXX) $(CPPFLAGS) $(CXXFLAGS) $(LDFLAGS) $< -o $@
 
 integrands/failing: integrands/failing.cpp Makefile
 	$(COMP)
@@ -56,9 +61,6 @@ integrands/turbulence3_hcubature_v: integrands/turbulence3.cpp Makefile
 integrands/turbulence3_pcubature_v: integrands/turbulence3.cpp Makefile
 	$(COMP) -DVECTORIZE -DPCUBATURE -lpcubature -lboost_program_options
 
-integrands/turbulence4: integrands/turbulence4.cpp Makefile
-	$(COMP) -I $(HOME)/ohjelmat/cxx-prettyprint -lboost_program_options
-
 c: clean
 clean:
 	rm -f $(PROGRAMS) tests/*out tests/*ok
@@ -67,9 +69,9 @@ t: test
 test: tests/2d_ok tests/3d_ok
 
 tests/2d_ok: hdintegrator.py integrands/N-sphere.py Makefile
-	@printf 'TEST N-sphere.py 2d... ' && $(MPIEXEC) -n 2 ./hdintegrator.py --integrand integrands/N-sphere.py --dimensions 1 > tests/2d_out
+	@printf 'TEST N-sphere.py 2d... ' && $(MPIEXEC) -n 2 ./hdintegrator.py --integrand integrands/N-sphere.py --dimensions 1 | $(PYTHON) -c "from sys import stdin; val,err,vol=stdin.read().split(); print('{:.12e} {:.4e} {:.12e}'.format(float(val),float(err),float(vol)))" > tests/2d_out
 	@$(DIFF) -q tests/2d_ref tests/2d_out && $(TOUCH) tests/2d_ok && echo PASSED
 
 tests/3d_ok: hdintegrator.py integrands/N-sphere.py Makefile
-	@printf 'TEST N-sphere.py 3d... ' && $(MPIEXEC) -n 2 ./hdintegrator.py --integrand integrands/N-sphere.py --dimensions 2 > tests/3d_out
+	@printf 'TEST N-sphere.py 3d... ' && $(MPIEXEC) -n 2 ./hdintegrator.py --integrand integrands/N-sphere.py --dimensions 2 | $(PYTHON) -c "from sys import stdin; val,err,vol=stdin.read().split(); print('{:.12e} {:.4e} {:.12e}'.format(float(val),float(err),float(vol)))" > tests/3d_out
 	@$(DIFF) -q tests/3d_ref tests/3d_out && $(TOUCH) tests/3d_ok && echo PASSED
