### 2017-11-28 a48acc6
Allow more customization of Makefile and update documentation.
