### 2017-11-28 iljah/hdintegrator#9 @iljah
Fixed by https://github.com/iljah/hdintegrator/commit/a48acc6b912c75c0735175e9932e37bf5e0245a9 , now you can give those to make, e.g.

    make CPPFLAGS=-I/opt/local/include ...
