### 2017-11-28 2d31189
Remove unmaintained integrand.
