### 2017-11-28 6c36d76
Rearrange install instructions.
