### 2017-11-28 iljah/hdintegrator#11 @iljah
Fixed by https://github.com/iljah/hdintegrator/commit/7065d95d8af7b0ed12b12232fc1099ec6bddfd09 and previous commits.
