### 2017-11-28 iljah/hdintegrator#10 @iljah
I wasn't using/maintaining integrand(s) requiring cubature so I removed them instead.
