### 2017-11-28 a6f5b63
Mention in install that boost program options is also required.
