### 2017-11-28 iljah/hdintegrator#12 @iljah
Boost program options, added to install docs in https://github.com/iljah/hdintegrator/commit/a6f5b63a1c4a0a045ce3c23ff8da14b9097b78df
