### 2017-11-28 #437 @iljah
Fixed issues 9 to 12, thanks!
