### 2017-11-28 iljah/hdintegrator#2 @iljah
Good suggestion, I was planning to start using markdown at some point.
