### 2017-12-10 iljah/hdintegrator#12 @marksantcroos
I had to change the Makefile and change `-lboost_program_options` into `-lboost_program_options-mt` to get it compiled. But this might be an issue of how boost is installed on my machine (through macports).
