### 2017-12-10 #437 @marksantcroos
I updated iljah/hdintegrator#10 and iljah/hdintegrator#12, and added a new iljah/hdintegrator#13.
