### 2017-12-11 3ddc7e2
Allow further customization of environment in makefile.
