### 2017-12-11 iljah/hdintegrator#10 @iljah
Fixed by https://github.com/iljah/hdintegrator/commit/3ddc7e2ec674499dc8ddb2c6add87a80da6d2c15 thanks.
