### 2017-12-11 iljah/hdintegrator#13 @iljah
Hmm your command worked for me, the result was:

    3.3696876509523746e-06 1.9440074486843323e-07 0.0

Please try the command with `--verbose` and attach the last 100 lines or so.
