### 2017-12-11 #437 @iljah
Thanks, fixed 10 and 12 but I need more info for 13 because the command that failed for you worked for me.
