### 2017-12-11 iljah/hdintegrator#10 @marksantcroos
Thanks, build works now!
