### 2017-12-11 iljah/hdintegrator#13 @marksantcroos
```
(ve) netbook:hdintegrator mark$ mpiexec -n 2 ./hdintegrator.py --integrand integrands/N-sphere --dimensions 15 --prerefine 200 --verbose
Starting with 2 processes
Traceback (most recent call last):
  File "./hdintegrator.py", line 356, in <module>
    split(choice(grid.get_cells()), 1, [randint(0, len(dimensions) - 1)], grid)
  File "/Users/mark/Desktop/review/ve/lib/python3.6/random.py", line 258, in choice
    return seq[i]
  File "/Users/mark/Desktop/review/ve/lib/python3.6/site-packages/networkx/classes/reportviews.py", line 178, in __getitem__
    return self._nodes[n]
KeyError: 0
Integrand initialized by rank 1
Rank 1 waiting for work
```
