### 2017-12-11 iljah/hdintegrator#13 @iljah
Ok something strange is happening on process 0 and I can't tell if it's a problem with hdintegrator or networkx. This is what I get:

    Starting with 2 processes
    Integrand initialized by rank 1
    Rank 1 waiting for work
    Grid initialized by rank 0 with 201 cells
    Number of work item slots: 1
    201 work left, 0 processing
    Sending cell 3292 for processing to rank 1

Try changing

    				split(random_cell, 1, random_ints, grid)

around line 360 in hdintegrator.py to

    				try:
    					random_cell = choice(grid.get_cells())
    					random_int = [randint(0, len(dimensions) - 1)]
    					split(random_cell, 1, random_int, grid)
    				except Exception as e:
    					print("Couldn't split cell", random_cell, 'along', random_int, 'because', e)
    					exit(1)

and let me know what that prints.
