### 2017-12-11 iljah/hdintegrator#13 @iljah
This is the exact place online: https://github.com/iljah/hdintegrator/blob/master/hdintegrator.py#L356 and oops, I put the new version of the `split` line as the old one. The link above is the correct place to change.
