### 2017-12-13 iljah/hdintegrator#14 @marksantcroos
tabs vs spaces
Do you have any reason to use tabs instead of spaces?

It gave me a hard time to make the changes you suggested on #13.

Also see https://www.python.org/dev/peps/pep-0008/#tabs-or-spaces
