### 2017-12-13 iljah/hdintegrator#14 @iljah
I'm using tabs so that the width of indentation can be customized without changing the code. Can you elaborate on why the changes in #13 were difficult? For me it was easy because the tabs were preserved so the replacement code worked as is.
