### 2017-12-13 iljah/hdintegrator#14 @marksantcroos
Because my main editor is configured confirming to pep8.
