### 2017-12-13 iljah/hdintegrator#13 @marksantcroos
See below. Note that the values used in the error message are uninitialised.

```
(ve) netbook:hdintegrator mark$ mpiexec -n 2 ./hdintegrator.py --integrand integrands/N-sphere --dimensions 15 --prerefine 200 --verbose
Starting with 2 processes
Traceback (most recent call last):
  File "./hdintegrator.py", line 358, in <module>
    random_cell = choice(grid.get_cells())
  File "/Users/mark/Desktop/review/ve/lib/python3.6/random.py", line 258, in choice
    return seq[i]
  File "/Users/mark/Desktop/review/ve/lib/python3.6/site-packages/networkx/classes/reportviews.py", line 178, in __getitem__
    return self._nodes[n]
KeyError: 0

Traceback (most recent call last):
  File "./hdintegrator.py", line 363, in <module>
    "Couldn't split cell", random_cell, 'along', random_int, 'because', e)
NameError: name 'random_cell' is not defined
Integrand initialized by rank 1
Rank 1 waiting for work
```
