### 2017-12-13 iljah/hdintegrator#14 @iljah
I didn't see copy & paste mentioned in pep8 but it does say that tabs should be used to remain consistent with code already indented with tabs. Do you mean that your configuration converted tabs to spaces automatically? Shouldn't that apply to pasted code also?

In any case the tabs vs. spaces question was closed on stackexchange as not constructive so there are probably valid arguments on both sides but for now I'll stick to tabs.
