### 2017-12-13 iljah/hdintegrator#13 @iljah
I meant that one line should be replaced with the entire try+except code above, so that the new file looks like this between lines 352..364:

    				c.set_extent(i, args.min_extent, args.max_extent)
    			grid = ndgrid(c)

    			for i in range(args.prerefine):
    				try:
    					random_cell = choice(grid.get_cells())
    					random_int = [randint(0, len(dimensions) - 1)]
    					split(random_cell, 1, random_int, grid)
    				except Exception as e:
    					print("Couldn't split cell", random_cell, 'along', random_int, 'because', e)
    					exit(1)
    			if args.verbose:
    				print('Grid initialized by rank', rank, 'with', len(grid.get_cells()), 'cells')

this way we should see if `split` is called with sane arguments and I should be able to reproduce the problem.
