### 2017-12-13 iljah/hdintegrator#13 @marksantcroos
Yes, I did that. If you look careful, you will see that random_cell and random_int will be uninitialised if the `choice` line fails (which it does).
