### 2017-12-13 iljah/hdintegrator#14 @marksantcroos
I can't force you, but as a piece of feedback, for me it would be a showstopper to contribute to your project.
