### 2017-12-13 iljah/hdintegrator#14 @iljah
Ok thanks for the input, I'll keep it in mind. Is this the case with any language or only python?
