### 2017-12-13 iljah/hdintegrator#13 @iljah
Oh right, so does `grid.get_cells()` return an empty list then in this case?
