### 2017-12-13 iljah/hdintegrator#14 @marksantcroos
With python in particular. The reference to "code that was already indented with tabs" referred to code already written at the time that pep8 was published.
