### 2017-12-13 iljah/hdintegrator#13 @marksantcroos
A NodeView object.
