### 2017-12-13 iljah/hdintegrator#13 @iljah
Looks like networkx api was changed from 1.x to 2.x. and I'm still using 1.x series. I've updated ndgrid and the corresponding submodule in hdintegrator, if you do a `git pull` followed by `git submodule update` it should work. If it doesn't and ndgrid in submodule dir looks like https://github.com/iljah/ndgrid/blob/master/source/ndgrid.py#L71 then something else might be the problem.
