### 2017-12-15 #437 @iljah
Glad to hear! I'll create a tag in the repository and update the pypi version.
