### 2017-12-15 cfcc6cf
Update codemeta and setup in sync.
