### 2017-12-15 #437 @danielskatz
@iljah - next we need the DOI to an archived version of the current software

And @marksantcroos - thanks! you've made a good contribution to the software, the paper, and the journal.
