### 2017-12-15 #437 @iljah
I've created a DOI for the archive at Zenodo: 10.5281/zenodo.1116815
