### 2017-12-15 #437 @danielskatz
@arfon - I think this is ready for you to accept
